Switched environment to PHP 7.0.33
Now in line with minimal requirement
